requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
djangorestframework==3.14.0
greenlet==2.0.2
iniconfig==2.0.0
kombu==5.3.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
